+++ openrem/remapp/extractors/rdsr.py
@@ -625,8 +625,19 @@ def _ctirradiationeventdata(dataset, ct, ch):  # TID 10013
         if cont.ConceptNameCodeSequence[0].CodeMeaning == 'Acquisition Protocol':
             event.acquisition_protocol = safe_strings(cont.TextValue, char_set=ch)
         elif cont.ConceptNameCodeSequence[0].CodeMeaning == 'Target Region':
-            event.target_region = get_or_create_cid(cont.ConceptCodeSequence[0].CodeValue,
-                                                    cont.ConceptCodeSequence[0].CodeMeaning)
+            try:
+                event.target_region = get_or_create_cid(cont.ConceptCodeSequence[0].CodeValue,
+                                                        cont.ConceptCodeSequence[0].CodeMeaning)
+            except AttributeError:
+                logger.info(u'Target Region ConceptNameCodeSequence exists, but no content. Study UID {0} from {1}, '
+                               u'{2}, {3}'.format(
+                                event.ct_radiation_dose.general_study_module_attributes.study_instance_uid,
+                                event.ct_radiation_dose.general_study_module_attributes.generalequipmentmoduleattr_set.get(
+                                    ).manufacturer,
+                                event.ct_radiation_dose.general_study_module_attributes.generalequipmentmoduleattr_set.get(
+                                    ).manufacturer_model_name,
+                                event.ct_radiation_dose.general_study_module_attributes.generalequipmentmoduleattr_set.get(
+                                    ).station_name))
         elif cont.ConceptNameCodeSequence[0].CodeMeaning == 'CT Acquisition Type':
             event.ct_acquisition_type = get_or_create_cid(cont.ConceptCodeSequence[0].CodeValue,
                                                           cont.ConceptCodeSequence[0].CodeMeaning)
@@ -944,15 +955,19 @@ def _generalstudymoduleattributes(dataset, g, ch):
     g.procedure_code_meaning = get_seq_code_meaning('ProcedureCodeSequence', dataset)
     g.requested_procedure_code_value = get_seq_code_value('RequestedProcedureCodeSequence', dataset)
     g.requested_procedure_code_meaning = get_seq_code_meaning('RequestedProcedureCodeSequence', dataset)
+    g.save()
+
     try:
-        if dataset.ContentTemplateSequence[0].TemplateIdentifier == '10001':
-            _projectionxrayradiationdose(dataset, g, 'projection', ch)
-        elif dataset.ContentTemplateSequence[0].TemplateIdentifier == '10011':
-            _projectionxrayradiationdose(dataset, g, 'ct', ch)
+        template_identifier = dataset.ContentTemplateSequence[0].TemplateIdentifier
     except AttributeError:
         logger.error(u"Study UID {0} of modality {1} has no template sequence - incomplete RDSR. Aborting.".format(
             g.study_instance_uid, get_value_kw("ManufacturerModelName", dataset)))
         g.delete()
+        return
+    if template_identifier == '10001':
+        _projectionxrayradiationdose(dataset, g, 'projection', ch)
+    elif template_identifier == '10011':
+        _projectionxrayradiationdose(dataset, g, 'ct', ch)
     g.save()
     if not g.requested_procedure_code_meaning:
         if (('RequestAttributesSequence' in dataset) and dataset[